+++ panda/board/obj/gitversion.h
@@ -1,1 +1,1 @@
-const uint8_t gitversion[] = "DEV-7a756b69-DEBUG";
+const uint8_t gitversion[] = "DEV-6afa1a46-DEBUG";