+++ selfdrive/controls/lib/desire_helper.py
@@ -76,6 +76,16 @@ class DesireHelper:
       elif self.lane_change_state == LaneChangeState.laneChangeStarting:
         # fade out over .5s
         self.lane_change_ll_prob = max(self.lane_change_ll_prob - 2 * DT_MDL, 0.0)
+        #stop if touching wheel or blinker in the other direction
+        if (
+          (carstate.rightBlinker and self.lane_change_direction == LaneChangeDirection.left)
+          or
+          (carstate.leftBlinker and self.lane_change_direction == LaneChangeDirection.right)
+          or
+          carstate.steeringPressed
+        ):
+          self.lane_change_direction = LaneChangeDirection.none
+          self.lane_change_state = LaneChangeState.off
 
         # 98% certainty
         if lane_change_prob < 0.02 and self.lane_change_ll_prob < 0.01: